refactor: move fragment up one level
