+++ libs/frontend/modules/tag/src/use-cases/tag.endpoints.graphql.gen.ts
@@ -1,8 +1,11 @@
 import * as Types from '@codelab/frontend/abstract/codegen'
 
-import { TagFragment, TagGraphFragment } from './Tag.fragment.graphql.gen'
+import { TagFragment, TagGraphFragment } from '../Tag.fragment.graphql.gen'
 import { gql } from '@apollo/client'
-import { TagFragmentDoc, TagGraphFragmentDoc } from './Tag.fragment.graphql.gen'
+import {
+  TagFragmentDoc,
+  TagGraphFragmentDoc,
+} from '../Tag.fragment.graphql.gen'
 import { api, GraphqlOperationOptions } from '@codelab/frontend/model/infra/api'
 export type CreateTagMutationVariables = Types.Exact<{
   input: Types.CreateTagInput
